+++ src/widgets/commands.ts
@@ -1,0 +1,53 @@
+// Copyright (c) Jeremy Tuloup
+// Distributed under the terms of the Modified BSD License.
+
+import { CommandRegistry } from '@phosphor/commands';
+
+import { ISerializers, WidgetModel } from '@jupyter-widgets/base';
+
+import { MODULE_NAME, MODULE_VERSION } from '../version';
+
+export class CommandRegistryModel extends WidgetModel {
+  defaults() {
+    return {
+      ...super.defaults(),
+      _model_name: CommandRegistryModel.model_name,
+      _model_module: CommandRegistryModel.model_module,
+      _model_module_version: CommandRegistryModel.model_module_version
+    };
+  }
+
+  initialize(attributes: any, options: any) {
+    this.commands = CommandRegistryModel._commands;
+    super.initialize(attributes, options);
+    this.on('msg:custom', this.onMessage.bind(this));
+
+    this.set('_commands', this.commands.listCommands());
+    this.save_changes();
+  }
+
+  private onMessage(msg: any) {
+    switch (msg.func) {
+      case 'execute':
+        const { command, args } = msg.payload;
+        void this.commands.execute(command, args);
+        break;
+      default:
+        break;
+    }
+  }
+
+  static serializers: ISerializers = {
+    ...WidgetModel.serializers
+  };
+
+  static model_name = 'CommandRegistryModel';
+  static model_module = MODULE_NAME;
+  static model_module_version = MODULE_VERSION;
+  static view_name: string = null;
+  static view_module: string = null;
+  static view_module_version = MODULE_VERSION;
+
+  private commands: CommandRegistry;
+  static _commands: CommandRegistry;
+}
